Refactored graphql endpoints to commonize logic
